Complete Basic Crawling

+++ search.php
@@ -4,44 +4,56 @@ require_once 'sql.php';
 
 $user_name = 'kenkoooo';
 
-$abcArray = getProblemArray ( '/abc[0-9]*/i', 2012 );
-$arcArray = getProblemArray ( '/arc[0-9]*/i', 2012 );
-$allArray = getProblemArray ( '/^(?!.*(abc|arc)).*$/', 2013 );
+$abcArray = getProblemArray ( '/abc[0-9]*/i', 2012, $user_name );
+$arcArray = getProblemArray ( '/arc[0-9]*/i', 2012, $user_name );
+$allArray = getProblemArray ( '/^(?!.*(abc|arc)).*$/', 2013, $user_name );
 
 include 'html.inc';
-function getProblemArray($pattern, $year) {
+function getProblemArray($pattern, $year, $user_name) {
 	// 正規表現にマッチするコンテストネームの問題集を返す
 	// $year以降のコンテストを返す
 	$array = array ();
 	$sql = new SQLConnect ();
 	$pull = $sql->pullContests ();
 	
+	// コンテスト情報を取得
 	foreach ( $pull as $element ) {
 		$name = $element ["name"];
 		$end = $element ["end"];
 		
-		if (preg_match ( $pattern, $element ["name"] ) && strtotime ( $end ) > strtotime ( $year . "/01/01" )) {
-			array_push ( $array, $element );
+		if (preg_match ( $pattern, $name ) && strtotime ( $end ) > strtotime ( $year . "/01/01" )) {
+			$array [$name] = $element;
 		}
 	}
 	
-	// // 列nameでソートしたい
-	// foreach ( $array as $key => $row ) {
-	// $n [$key] = $row ["name"];
-	// }
-	// array_multisort ( $n, SORT_ASC, $array );
-	
-	for($i = 0; $i < count ( $array ); $i ++) {
-		$problems = $sql->getProblems ( $array [$i] ["id"] );
-		$array [$i] ["problems"] = array ();
+	// 問題情報を取得
+	foreach ( $array as $key => $contest ) {
+		$problems = $sql->getProblems ( $contest ["id"] );
 		foreach ( $problems as $p ) {
-			array_push ( $array [$i] ["problems"], $p );
+			$problem_name = $p ["name"];
+			$array [$key] ["problems"] [$problem_name] = $p;
+			$array [$key] ["problems"] [$problem_name] ["solved"] = FALSE;
 		}
 	}
+	
+	$solved = $sql->getSolved ( $user_name );
+	foreach ( $solved as $sol ) {
+		$contest_name = $sol ["contest_name"];
+		$problem_name = $sol ["problem_name"];
+		if (array_key_exists ( $contest_name, $array )) {
+			$array [$contest_name] ["problems"] [$problem_name] ["solved"] = TRUE;
+		}
+	}
+	
 	return $array;
 }
 function listupARC($array) {
 	foreach ( $array as $contest ) {
+		if (! array_key_exists ( "problems", $contest )) {
+			// 問題が存在しなければスルー
+			continue;
+		}
+		
 		echo '<tr>';
 		
 		$contest_name = $contest ["name"];
@@ -57,7 +69,11 @@ function listupARC($array) {
 			
 			echo "<td><a href='http://$contest_name.contest.atcoder.jp/tasks/$contest_problem_name'>";
 			echo $contest_problem_title;
-			echo "</a></td>";
+			echo "</a>";
+			if ($contest_problem ["solved"]) {
+				echo "<br><span class='label label-success'>AC</span>";
+			}
+			echo "</td>";
 		}
 		
 		echo '</tr>';
